@@ -2,8 +2,8 @@ package edu.cs4730.nearbymessagesdemo;
 
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -56,7 +56,8 @@ public class PublishFragment extends Fragment {
 
     private void publish(String message) {
         logthis("Publishing message: " + message);
-        mActiveMessage = new Message("Hello World".getBytes());
+       // mActiveMessage = new Message("Hello World".getBytes());
+        mActiveMessage = new Message(message.getBytes());
 
         PublishOptions options = new PublishOptions.Builder()
             .setStrategy(PUB_SUB_STRATEGY)
